+++ spec/directions/bake_iframes/v1_spec.rb
@@ -74,7 +74,7 @@ RSpec.describe Kitchen::Directions::BakeIframes::V1 do
       <<~HTML
               <div data-alt="atoms_isotopes" data-type="media" id="auto_78a58d46-6a29-4972-b848-385be0c8057f_11">
                 <div class="os-has-iframe os-has-link" data-type="alternatives">
-                  <a class="os-is-link" href="../../media/Interactive/Comp_Figure 18_18.5_020/index.html" target="_window">Click to view content</a>
+                  <a class="os-is-link" href="../../media/Interactive/Comp_Figure 18_18.5_020/index.html" target="_window">Click to view content ("../../media/Interactive/Comp_Figure 18_18.5_020/index.html")</a>
                   <iframe class="os-is-iframe" height="371.4" src="../../media/Interactive/Comp_Figure 18_18.5_020/index.html" width="660">
         <!-- no-selfclose -->
                   </iframe>
@@ -90,7 +90,7 @@ RSpec.describe Kitchen::Directions::BakeIframes::V1 do
       <<~HTML
             <div class="interactive" data-has-label="true" data-label="" data-type="note" id="iframenote">
                 <div data-alt="atoms_isotopes" data-type="media">
-                  <div class="os-has-iframe os-has-link" data-type="alternatives"><a class="os-is-link" href="https://openstax.org/l/atoms_isotopes" target="_window">Click to view content</a>
+                  <div class="os-has-iframe os-has-link" data-type="alternatives"><a class="os-is-link" href="https://openstax.org/l/atoms_isotopes" target="_window">Click to view content ("https://openstax.org/l/atoms_isotopes")</a>
                     <iframe class="os-is-iframe" height="371.4" src="https://openstax.org/l/atoms_isotopes" width="660"><!-- no-selfclose -->
                     </iframe>
                   </div>
@@ -109,7 +109,7 @@ RSpec.describe Kitchen::Directions::BakeIframes::V1 do
               <ul>
                 <li>1: The evolution from fish to earliest tetrapod<span data-type="newline"><br /></span>
               <div data-alt="tetrapod_evol1" data-type="media">
-                <div class="os-has-iframe os-has-link" data-type="alternatives"><a class="os-is-link" href="url1" target="_window">Click to view content</a>
+                <div class="os-has-iframe os-has-link" data-type="alternatives"><a class="os-is-link" href="url1" target="_window">Click to view content ("url1")</a>
                   <iframe class="os-is-iframe" height="371.4" src="url1" width="660"><!-- no-selfclose -->
                   </iframe>
                 </div>
@@ -117,7 +117,7 @@ RSpec.describe Kitchen::Directions::BakeIframes::V1 do
             </li>
                 <li>2: The discovery of coelacanth and <em data-effect="italics">Acanthostega</em> fossils<span data-type="newline"><br /></span>
               <div data-alt="tetrapod_evol2" data-type="media">
-                <div class="os-has-iframe os-has-link" data-type="alternatives"><a class="os-is-link" href="url2" target="_window">Click to view content</a>
+                <div class="os-has-iframe os-has-link" data-type="alternatives"><a class="os-is-link" href="url2" target="_window">Click to view content ("url2")</a>
                   <iframe class="os-is-iframe" height="371.4" src="url2" width="660"><!-- no-selfclose -->
                   </iframe>
                 </div>